+++ src/main/java/com/levelUpToast/levelUpToast/rest/img/ImgController.java
@@ -38,7 +38,7 @@ public class ImgController {
             log.info("[ImgController log] : 이미지 저장 성공 originalImgName = {}, uploadImgName = {}",imgItem.getUploadFileName(),imgItem.getStoreFileName());
             return new ImgSaveResponseForm(-1, "img 이름 : " + imgItem.getUploadFileName() + " 저장완료", data);
         } catch (Exception e) {
-            log.info("ex ={}",e);
+            log.info("Ex = {}", e.toString());
             return new ImgSaveResponseForm(112, e.getMessage(), null);
         }
